Removed 64 bit integers in shader

+++ src/renderer/scene.py
@@ -29,10 +29,10 @@ class Material(ct.Structure):
 
 class Mesh(ct.Structure):
     _fields_ = [
-        ("startingVertex", ct.c_uint64),
-        ("numTriangles", ct.c_uint64),
-        ("materialID", ct.c_uint64),
-        ("objectID", ct.c_uint64),
+        ("startingVertex", ct.c_uint32),
+        ("numTriangles", ct.c_uint32),
+        ("materialID", ct.c_uint32),
+        ("objectID", ct.c_uint32),
     ]
